requirements.txt
python-snap7>=2.0.0
paho-mqtt>=1.6.0 